The Agent Host supervisor was locald's last file over 600 lines
1,564 of them: whether the sidecar should be running, starting and
ending the process, its log, the CLI calls that pair and unpair, and the
status the shell reads. Six modules -- `lifecycle`, `process`, `cli`,
`pairing`, `status` -- with the guards split the same way.

The guard on `run_cli` was reading the old file by name and sliced two
thousand *bytes* forward from it. It reads the module that owns
`run_cli` now, and takes two thousand characters: these sources are as
much prose as code, and a byte offset that lands inside an em dash
panics.

Nothing in `locald` is over 600 lines any more, and nothing in the
desktop shell or the local runtime is either. What is left in the
baseline is the agent host, and that is its own change.
