+++ desktop/scripts/file-size-baseline.json
@@ -8,6 +8,5 @@
   "desktop/agent-host/src/runtime.rs": 5126,
   "desktop/agent-host/tests/lemma_mcp_e2e.rs": 699,
   "desktop/agent-host/tests/real_harness_e2e.rs": 1202,
-  "desktop/agent-host/tests/support/mod.rs": 1417,
-  "desktop/locald/src/agent_host.rs": 1564
+  "desktop/agent-host/tests/support/mod.rs": 1417
 }
